nice UI

+++ appLogic.R
@@ -32,25 +32,40 @@ observeEvent(input$createReport, {
         app_key <- session$userData$appKey
         app_secret <- session$userData$appSecret
         private_key <- as.character(session$userData$keyItems)[3]
-        taskStr = paste0(
-                "docker run -i --rm ",
-                "--env-file <(env | grep MAIL) ",
-                "--env-file <(env | grep QUEUE) ",
-                "--link $DOCKER_LINK_MQ ",
-                "oydeu/srv-report /bin/run.sh '{",
-                '"pia_url":"', pia_url, '",',
-                '"app_key":"', app_key, '",',
-                '"app_secret":"', app_secret, '",',
-                '"private_key":"', private_key, '"}', "'")
-        # oydLog(paste("task:", taskStr))
-        # system(taskStr)
-        
-        # should write task into Tasks queue - but does not work yet!!!
-        app <- oydapp::setupApp(pia_url, app_key, app_secret, '')
-        oydapp::createTask(
-                app,
-                'oyd.report.temp', # identifier
-                taskStr,           # command
-                'delete'           # schedule
-        )
+        if(as.character(private_key) != '' && !is.na(private_key)){
+                shinyWidgets::sendSweetAlert(
+                        session = session,
+                        title = tr('createReportTitle'),
+                        text = tr('createReportInfo'),
+                        type = "success"
+                )
+                taskStr = paste0(
+                        "docker run -i --rm ",
+                        "--env-file <(env | grep MAIL) ",
+                        "--env-file <(env | grep QUEUE) ",
+                        "--link $DOCKER_LINK_MQ ",
+                        "oydeu/srv-report /bin/run.sh '{",
+                        '"pia_url":"', pia_url, '",',
+                        '"app_key":"', app_key, '",',
+                        '"app_secret":"', app_secret, '",',
+                        '"private_key":"', private_key, '"}', "'")
+                # oydLog(paste("task:", taskStr))
+                # system(taskStr)
+                
+                # should write task into Tasks queue - but does not work yet!!!
+                app <- oydapp::setupApp(pia_url, app_key, app_secret, '')
+                oydapp::createTask(
+                        app,
+                        'oyd.report.temp', # identifier
+                        taskStr,           # command
+                        'delete'           # schedule
+                )
+        } else {
+                shinyWidgets::sendSweetAlert(
+                        session = session,
+                        title = tr('missingKey'),
+                        text = tr('missingKeyInfo'),
+                        type = "error"
+                )
+        }
 })
